requirements.txt
telethon
cryptg 
aiohttp
python-dotenv